Restructre hpu pytorch lightning code
Align the code hierarch similar to root and enable tests
