+++ examples/pytorch/mnist_sample.py
@@ -11,14 +11,24 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import torch
-from jsonargparse import lazy_instance
-from pytorch_lightning import LightningModule
-from pytorch_lightning.cli import LightningCLI
-from pytorch_lightning.demos.mnist_datamodule import MNISTDataModule
-from torch.nn import functional as F  # noqa: N812
 
-from lightning_habana.plugins.precision import HPUPrecisionPlugin
+import torch
+from torch.nn import functional as F  # noqa: N812
+from jsonargparse import lazy_instance
+
+from lightning_utilities import module_available
+
+if module_available("lightning"):
+    from lightning.pytorch import LightningModule
+    from lightning.pytorch.cli import LightningCLI
+    from lightning.pytorch.demos.mnist_datamodule import MNISTDataModule
+elif module_available("pytorch_lightning"):
+    from pytorch_lightning import LightningModule
+    from pytorch_lightning.cli import LightningCLI
+    from pytorch_lightning.demos.mnist_datamodule import MNISTDataModule
+
+
+from lightning_habana.pytorch.plugins.precision import HPUPrecisionPlugin
 
 
 class LitClassifier(LightningModule):
@@ -31,7 +41,8 @@ class LitClassifier(LightningModule):
 
     def training_step(self, batch, batch_idx):
         x, y = batch
-        return F.cross_entropy(self(x), y)
+        loss = F.cross_entropy(self(x), y)
+        return loss
 
     def validation_step(self, batch, batch_idx):
         x, y = batch
@@ -47,7 +58,8 @@ class LitClassifier(LightningModule):
 
     @staticmethod
     def accuracy(logits, y):
-        return torch.sum(torch.eq(torch.argmax(logits, -1), y).to(torch.float32)) / len(y)
+        acc = torch.sum(torch.eq(torch.argmax(logits, -1), y).to(torch.float32)) / len(y)
+        return acc
 
     def configure_optimizers(self):
         return torch.optim.Adam(self.parameters(), lr=0.02)
@@ -61,7 +73,7 @@ if __name__ == "__main__":
             "accelerator": "hpu",
             "devices": 1,
             "max_epochs": 1,
-            "plugins": lazy_instance(HPUPrecisionPlugin, precision="16-mixed"),
+            "plugins": lazy_instance(HPUPrecisionPlugin, precision="bf16-mixed"),
         },
         run=False,
         save_config_kwargs={"overwrite": True},
