+++ examples/pytorch/ops_bf16_mnist.txt
