+++ examples/pytorch/ops_fp32_mnist.txt
