+++ src/lightning_habana/__about__.py
@@ -1,10 +1,10 @@
-__version__ = "0.1.0rc1"
+__version__ = "1.10.0.0"
 __author__ = "Lightning-AI et al."
 __author_email__ = "name@lightning.ai"
 __license__ = "Apache-2.0"
 __copyright__ = f"Copyright (c) 2020-2022, {__author__}."
 __homepage__ = "https://github.com/Lightning-AI/lightning-habana"
-__docs__ = "PyTorch Lightning Sample project."
+__docs__ = "Lightning suport for Intel Habana accelerators"
 
 __all__ = [
     "__author__",
