+++ src/lightning_habana/pytorch/accelerator/__init__.py
@@ -12,7 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from lightning_habana.__about__ import *  # noqa: E402, F401, F403
-from lightning_habana.pytorch import *  # noqa: E402
-from lightning_habana.fabric import *  # noqa: E402
-from lightning_habana.utils import *  # noqa: E402
+from lightning_habana.pytorch.accelerator.hpu import HPUAccelerator  # noqa: E402
+__all__ = ["HPUAccelerator"]
