+++ src/lightning_habana/utils/imports.py
@@ -12,7 +12,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from lightning_habana.__about__ import *  # noqa: E402, F401, F403
-from lightning_habana.pytorch import *  # noqa: E402
-from lightning_habana.fabric import *  # noqa: E402
-from lightning_habana.utils import *  # noqa: E402
+import os
+import operator
+import sys
+
+import torch
+
+_PACKAGE_ROOT = os.path.dirname(__file__)
+_PROJECT_ROOT = os.path.dirname(_PACKAGE_ROOT)
+
+from lightning_utilities.core.imports import compare_version, package_available, RequirementCache  # noqa: E402
+
+_HABANA_FRAMEWORK_AVAILABLE = package_available("habana_frameworks")
+if _HABANA_FRAMEWORK_AVAILABLE:
+    from habana_frameworks.torch.utils.library_loader import is_habana_available
+
+    _HPU_AVAILABLE: bool = is_habana_available()
+else:
+    _HPU_AVAILABLE = False
+
+_TORCH_LESSER_EQUAL_1_13_1 = compare_version("torch", operator.le, "1.13.1")
