+++ tests/tests_pytorch/__init__.py
@@ -1,4 +1,4 @@
-# Copyright The Lightning AI team.
+# Copyright (c) 2023 Habana Labs, Ltd. an Intel Company
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from lightning_habana.strategies.parallel import HPUParallelStrategy
-from lightning_habana.strategies.single import SingleHPUStrategy
 
-__all__ = ["HPUParallelStrategy", "SingleHPUStrategy"]
+from lightning_habana.utils.imports import _HPU_AVAILABLE  # noqa: F401
+
+assert _HPU_AVAILABLE
