+++ tests/tests_pytorch/ops_bf16.txt
@@ -1,0 +1,2 @@
+linear
+relu
