+++ tests/tests_pytorch/ops_fp32.txt
@@ -1,0 +1,1 @@
+cross_entropy
